refactor(artifact): 뷰어 헬퍼를 service로 추출 (파일 크기 가드 600 준수)
artifacts.routes.ts 604→565줄. resolveAuthorLabel/exportPublicationViewer/
composeShareUrl 을 artifact-viewer-service 로 이동(동작 동일).
